Separate token and secret
We don't have any need for the token at this point since we're
validating a OAuth request that doesn't have one yet, but relying on
consumers of this function to remember to append an empty & is not a
good idea (since I forgot the first time). Instead I've added a token
parameter which can be passed in blank.