Launch for LTI v1.3
This new version of the LTI standard layers openidconnect and JSON
signature verification on top of the data we were expecting to get in an
LTI launch request. I've attempted to layer the new stuff in so that we
can support both protocols simultaneously for a transitionary period.
This required doing some interesting type discrimination stuff in
Typescript all of which will disappear when we only have a single
standard to support.

In addition to new data in the launch request we also have to handle a
request to a specific login URL as well.