Handle LTI 1.3 Events
Instead of sending data in the body the 1.3 LTI sends everything as a JWT, we need some of this information later to load our school configuration so we decode the payload and add it to our events. By using a discriminator property (ltiVersion) we can return two types of events and keep track of them later.

Because the jose library is provided as esm modules we have to mock it at the very top level and then access and use that mock in a slightly over complicated way, but it works.