Add LTI 1.3 Validation and Extraction Logic
These need more tests, but working for now. We use DynamoDB to store the state and the nonce (and the ttl option to ensure they get destroyed) and we pull the keyserver and use the jose library to validate the JWT we get back. Then pull the data we expect from the payload based on school configuration.